+++ src/main/java/org/dromara/easyai/naturalLanguage/languageCreator/CreatorWord.java
@@ -8,13 +8,13 @@ import java.util.List;
 
 public class CreatorWord implements OutBack {
     private int id;
-    private double maxOut = -2;
+    private float maxOut = -2;
 
     public int getId() {
         return id;
     }
 
-    public double getMaxOut() {
+    public float getMaxOut() {
         return maxOut;
     }
 
@@ -23,7 +23,7 @@ public class CreatorWord implements OutBack {
     }
 
     @Override
-    public void getBack(double out, int id, long eventId) {
+    public void getBack(float out, int id, long eventId) {
         if (out > maxOut) {
             maxOut = out;
             this.id = id;
@@ -31,7 +31,7 @@ public class CreatorWord implements OutBack {
     }
 
     @Override
-    public void getSoftMaxBack(long eventId, List<Double> softMax) {
+    public void getSoftMaxBack(long eventId, List<Float> softMax) {
 
     }
 
@@ -48,7 +48,7 @@ public class CreatorWord implements OutBack {
     }
 
     @Override
-    public void getWordVector(int id, double w) {
+    public void getWordVector(int id, float w) {
 
     }
 }
